separate renderer types
