+++ SnakeGame/rendererFactories/index.mjs
@@ -1,0 +1,2 @@
+export { default as canvas } from './canvas.mjs';
+export { default as element } from './element.mjs';
