+++ combineRenderers.mjs
@@ -1,0 +1,5 @@
+export default (...renderers) => (events) => {
+  for(const renderer of renderers){
+    renderer(events);
+  }
+}
